+++ lib/cache/web_cache.dart
@@ -1,3 +1,6 @@
+// ignore_for_file: curly_braces_in_flow_control_structures
+
+import 'dart:async';
 import 'dart:convert';
 // ignore: avoid_web_libraries_in_flutter
 import 'dart:html' as html;
@@ -6,47 +9,80 @@ import 'dart:typed_data';
 import 'package:flutter/services.dart';
 
 import './interface.dart';
+import '../image_info_data/image_info_data.dart';
 
 ImageCacheManger getInstance() => const _WebImageDataBase();
 
 class _WebImageDataBase extends ImageCacheManger {
   static final html.Storage caches = html.window.localStorage;
 
-  static late final List<String> sortedKeys;
+  static late final Map<String, Map<String, dynamic>> fileContent;
+
+  static late final bool _disableWebCache;
 
   const _WebImageDataBase();
 
   @override
-  Future<void> init() async {
-    sortedKeys = caches.keys.toList()
-      ..sort((a, b) => (json.decode(caches[a]!)['time']).compareTo(
-            json.decode(caches[b]!)['time'],
-          ));
+  Future<void> init(final bool enableWebCache) async {
+    if (!enableWebCache) {
+      _disableWebCache = true;
+      fileContent = {};
+      return;
+    }
+
+    _disableWebCache = false;
+
+    if (caches.keys.isEmpty) {
+      fileContent = {};
+      return;
+    }
+
+    for (final key in caches.keys)
+      if (_webKeyReg.hasMatch(key)) _sortedWebKeys.add(key);
+
+    fileContent = Map.from(
+      json.decode(_sortedWebKeys.join().replaceAll('}{', ',')),
+    );
+
+    _sortedWebKeys.sort((a, b) => (json.decode(caches[a]!)['time']).compareTo(
+          json.decode(caches[b]!)['time'],
+        ));
   }
 
   @override
-  Future<void> addNew({required String key, required Uint8List bytes}) async {
-    sortedKeys.add(key);
+  Future<void> addNew(final ImageInfoData imageInfo) async {
+    if (_disableWebCache || fileContent[imageInfo.key] != null) return;
+
+    fileContent.putIfAbsent(imageInfo.key, () => imageInfo.sizeToMap());
 
     try {
-      _add(key, bytes);
+      _add(imageInfo);
     } on html.DomException catch (e) {
       if (e.message?.contains('exceeded the quota') ?? false) {
         _removeOld5();
-        _add(key, bytes);
+        _add(imageInfo);
       }
     }
   }
 
   @override
+  ImageInfoData? getImageInfo(final String key) {
+    final data = fileContent[key];
+
+    if (data == null) return null;
+
+    return ImageInfoData.fromMap(data, key);
+  }
+
+  @override
   Future<Uint8List?> getBytes(String key) async {
-    if (caches.containsKey(key)) {
-      return Uint8List.fromList(
-        List<int>.from(json.decode(caches[key]!)['bytes']),
-      );
-    } else {
-      return null;
-    }
+    final imageInfo = getImageInfo(key);
+
+    if (imageInfo == null) return null;
+
+    return Uint8List.fromList(
+      List<int>.from(json.decode(caches[imageInfo.webKey()]!)['bytes']),
+    );
   }
 
   @override
@@ -57,27 +93,51 @@ class _WebImageDataBase extends ImageCacheManger {
 
   @override
   Future<void> clearCache() async {
-    sortedKeys.clear();
-    caches.clear();
+    fileContent.clear();
+
+    for (final key in _sortedWebKeys) {
+      caches.remove(key);
+      _sortedWebKeys.remove(key);
+    }
   }
 
-  void _add(final String key, final Uint8List bytes) {
+  static final _webKeyReg = RegExp(
+    r'{"http(s{0,1}).*?(width":\d+,"height":\d+})}',
+  );
+
+  static final List<String> _sortedWebKeys = [];
+
+  static void _removeOld5() {
+    final old5 = _sortedWebKeys.sublist(
+      0,
+      _sortedWebKeys.length > 5 ? 5 : _sortedWebKeys.length,
+    );
+
+    for (final key in old5) {
+      caches.remove(key);
+      _sortedWebKeys.remove(key);
+    }
+  }
+
+  static void _add(final ImageInfoData imageInfo) {
     try {
+      final key = imageInfo.webKey();
+
       caches.putIfAbsent(
         key,
         () => json.encode({
-          'bytes': bytes.toList(),
+          'bytes': imageInfo.memoryImage!.bytes.toList(),
           'time': DateTime.now().millisecondsSinceEpoch,
         }),
       );
+
+      _sortedWebKeys.add(key);
     } catch (e) {
       rethrow;
     }
   }
+}
 
-  void _removeOld5() {
-    final last5 = sortedKeys.sublist(0, 5);
-    caches.removeWhere((key, value) => last5.contains(key));
-    sortedKeys.removeRange(0, 5);
-  }
+extension on ImageInfoData {
+  String webKey() => json.encode({key: sizeToMap()});
 }